fix(rust): widen NumaNode.total_ram_mb to u64 to avoid truncation (C1)
total_ram_mb was u32, capping at ~4.29 TB per node: a larger node's
MemTotal (u64 kB / 1024) was silently truncated by the `as u32` cast
rather than reported or rejected. Make the field u64, matching the proto's
memory_mib, which also removes the cast in parse_node and the widening cast
in service.rs.

+++ rust/crates/supervisor-daemon/src/service.rs
@@ -241,7 +241,7 @@ fn numa_nodes_proto(topology: &crate::numa::NumaTopology) -> Vec<pb::NumaNode> {
         .map(|node| pb::NumaNode {
             index: node.id,
             cpu_count: node.cpus.len() as u32,
-            memory_mib: node.total_ram_mb as u64,
+            memory_mib: node.total_ram_mb,
         })
         .collect()
 }
